backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic-settings==2.3.0
httpx==0.27.0
openai==1.40.0
pillow==10.4.0
pdfplumber==0.11.4
python-multipart==0.0.9
pytest==8.3.0
pytest-asyncio==0.23.8
